Shift disabled logic to hooks to improve readability
