+++ src/commons/application/ApplicationWrapper.tsx
@@ -1,5 +1,4 @@
-import moment from 'moment';
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useMemo } from 'react';
 import { useDispatch } from 'react-redux';
 import { RouterProvider } from 'react-router';
 import { createBrowserRouter } from 'react-router-dom';
@@ -10,7 +9,7 @@ import {
   playgroundOnlyRouterConfig
 } from '../../routes/routerConfig';
 import Constants from '../utils/Constants';
-import { useSession } from '../utils/Hooks';
+import { useDisabled, useSession } from '../utils/Hooks';
 import { updateReactRouter } from './actions/CommonsActions';
 
 /**
@@ -24,33 +23,11 @@ import { updateReactRouter } from './actions/CommonsActions';
 const ApplicationWrapper: React.FC = () => {
   const dispatch = useDispatch();
   const { isLoggedIn, role, name, courseId } = useSession();
-
-  // Used in determining the disabled state of any type of Source Academy deployment (e.g. during exams)
-  const intervalId = useRef<number | undefined>(undefined);
-  const [isDisabled, setIsDisabled] = useState(computeDisabledState());
-
-  useEffect(() => {
-    if (Constants.disablePeriods.length > 0) {
-      intervalId.current = window.setInterval(() => {
-        const disabled = computeDisabledState();
-        if (isDisabled !== disabled) {
-          setIsDisabled(disabled);
-        }
-      }, 5000);
-    }
-
-    return () => {
-      if (intervalId.current) {
-        window.clearInterval(intervalId.current);
-      }
-    };
-  }, [isDisabled]);
+  const { disabledReason, isDisabledEffective } = useDisabled();
 
   const router = useMemo(() => {
-    const isDisabledEffective = !['staff', 'admin'].includes(role!) && isDisabled;
-
     const routerConfig = isDisabledEffective
-      ? getDisabledRouterConfig(isDisabled)
+      ? getDisabledRouterConfig(disabledReason)
       : Constants.playgroundOnly
       ? playgroundOnlyRouterConfig
       : getFullAcademyRouterConfig({
@@ -64,19 +41,9 @@ const ApplicationWrapper: React.FC = () => {
     dispatch(updateReactRouter(r));
 
     return r;
-  }, [isLoggedIn, isDisabled, role, name, courseId, dispatch]);
+  }, [disabledReason, isDisabledEffective, isLoggedIn, role, name, courseId, dispatch]);
 
   return <RouterProvider router={router} />;
 };
-
-function computeDisabledState() {
-  const now = moment();
-  for (const { start, end, reason } of Constants.disablePeriods) {
-    if (start.isBefore(now) && end.isAfter(now)) {
-      return reason || true;
-    }
-  }
-  return false;
-}
 
 export default ApplicationWrapper;
